# Improve trend unquarantine performance